pyproject.toml
[project]
name = "chatapp-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.63.0",
    "python-dotenv>=1.0.1",
    "reflex>=0.7.0",
]
